revert changes in requirements.txt file
This has beed done based on rohan's feedback, see:

https://github.com/Tendrl/alerting/pull/25#pullrequestreview-16714986

In the current state, requirements.txt handing is broken anyway, so it
make sense to fix/change in in a different pull request. In other words,
the requirements.txt file should be removed, but we will handle that
later, in a different commit. Moreover reverting the change prevents a
merge conflict.

tendrl-bug-id: Tendrl/alerting#24